+++ src/twoweeks/entities/PlayerCarClientAvatar.java
@@ -10,7 +10,7 @@ import twoweeks.models.CarAvatarModel;
 
 public class PlayerCarClientAvatar extends AbstractClientAvatar {
 	
-	public PlayerCarClientAvatar(AbstractGameClient _module, int _playerID, IInputDevice _input, Camera _cam, int eid, float x, float y, float z, int side) {
+	public PlayerCarClientAvatar(AbstractGameClient _module, int _playerID, IInputDevice _input, Camera _cam, int eid, float x, float y, float z, byte side) {
 		super(_module, TwoWeeksClientEntityCreator.CAR_AVATAR, _playerID, _input, _cam, eid, x, y, z, side, new CarAvatarModel(_module.getAssetManager()), null);
 		
 		//this.playerGeometry.setShadowMode(ShadowMode.Off);
